working without numerical property - validation is done manually and loading works
